Consolidate cache ownership
