+++ crates/unpack_core/src/compilation.rs
@@ -8,7 +8,7 @@ use crate::{
     Asset, ChunkGraph, CompilerOptions, Error, InfrastructureLogEvent, InfrastructureLogLevel,
     ModuleGraph, ModuleHandle, Result, UnpackResolver,
     build_chunk_graph::build_chunk_graph,
-    cache::BuildCache,
+    cache::Cache,
     code_generation::{self, CodeGenerationResults, RenderManifest},
     id_assignment::{assign_chunk_render_ids, assign_module_render_ids},
     make::{self, MakeState},
@@ -23,7 +23,7 @@ pub(crate) use hooks::CompilationHookSet;
 pub struct Compilation {
     options: CompilerOptions,
     resolver: UnpackResolver,
-    build_cache: BuildCache,
+    cache: Cache,
     module_graph: ModuleGraph,
     chunk_graph: ChunkGraph,
     render_ids_assigned: bool,
@@ -42,14 +42,14 @@ impl Compilation {
     pub(crate) fn new(
         options: CompilerOptions,
         resolver: UnpackResolver,
-        build_cache: BuildCache,
+        cache: Cache,
         hooks: CompilationHookSet,
     ) -> Self {
         let file_system_info = FileSystemInfo::from_snapshot_options(&options.snapshot);
         Self {
             options,
             resolver,
-            build_cache,
+            cache,
             module_graph: ModuleGraph::default(),
             chunk_graph: ChunkGraph::default(),
             render_ids_assigned: false,
@@ -123,7 +123,7 @@ impl Compilation {
             let result = make::run(
                 &self.options,
                 self.resolver.clone(),
-                self.build_cache.clone(),
+                self.cache.clone(),
                 self.file_system_info.clone(),
                 Arc::clone(&state),
             )
@@ -233,7 +233,7 @@ impl Compilation {
     pub fn render_assets(&mut self) {
         self.assets = code_generation::render_assets(
             &self.options,
-            &self.build_cache,
+            &self.cache,
             self.asset_render_manifest
                 .as_ref()
                 .expect("render manifest should exist before Asset rendering"),
@@ -253,7 +253,7 @@ impl Compilation {
         let outcome = code_generation::generate_code_cached(
             &self.module_graph,
             &self.chunk_graph,
-            &self.build_cache,
+            &self.cache,
         );
         self.errors.extend(outcome.errors);
         self.code_generation_results = Some(outcome.results);
