+++ crates/unpack_core/src/compiler.rs
@@ -8,7 +8,7 @@ use std::{
 
 use crate::{
     CacheOptions, Compilation, CompilationHooks, InfrastructureLoggingOptions, LoaderRunner,
-    ModuleRule, ResolveOptions, Result, SnapshotOptions, UnpackResolver, cache::BuildCache,
+    ModuleRule, ResolveOptions, Result, SnapshotOptions, UnpackResolver, cache::Cache,
     compilation::CompilationHookSet, flag_dependency_exports_plugin::FlagDependencyExportsPlugin,
     flag_dependency_usage_plugin::FlagDependencyUsagePlugin,
     optimize::side_effects_flag_plugin::SideEffectsFlagPlugin,
@@ -84,7 +84,7 @@ impl CompilerOptions {
 #[derive(Debug, Clone)]
 pub struct Compiler {
     options: CompilerOptions,
-    build_cache: BuildCache,
+    cache: Cache,
     cache_lifecycle: Arc<CacheLifecycle>,
     hooks: CompilerHookSet,
 }
@@ -200,17 +200,17 @@ struct CacheLifecycleState {
 
 #[derive(Debug)]
 struct CacheLifecycle {
-    build_cache: BuildCache,
+    cache: Cache,
     timeouts: CacheIdleTimeouts,
     state: Mutex<CacheLifecycleState>,
     changed: tokio::sync::watch::Sender<u64>,
 }
 
 impl CacheLifecycle {
-    fn new(build_cache: BuildCache, options: &CacheOptions) -> Arc<Self> {
+    fn new(cache: Cache, options: &CacheOptions) -> Arc<Self> {
         let (changed, _) = tokio::sync::watch::channel(0);
         Arc::new(Self {
-            build_cache,
+            cache,
             timeouts: CacheIdleTimeouts::from_options(options),
             state: Mutex::new(CacheLifecycleState {
                 activity: CacheActivity::Ready,
@@ -271,11 +271,10 @@ impl CacheLifecycle {
                 return;
             } else {
                 state.activity = CacheActivity::Idle;
-                if self.build_cache.pending_generation().is_some() {
+                if self.cache.pending_generation().is_some() {
                     let now = tokio::time::Instant::now();
                     state.ordinary_deadline = Some(now + self.timeouts.ordinary);
-                    if self.build_cache.initial_store_pending()
-                        && state.initial_store_deadline.is_none()
+                    if self.cache.initial_store_pending() && state.initial_store_deadline.is_none()
                     {
                         state.initial_store_deadline = Some(now + self.timeouts.initial_store);
                     }
@@ -309,7 +308,7 @@ impl CacheLifecycle {
             {
                 return;
             }
-            let Some(target_generation) = self.build_cache.pending_generation() else {
+            let Some(target_generation) = self.cache.pending_generation() else {
                 state.initial_store_deadline = None;
                 state.ordinary_deadline = None;
                 state.large_change_deadline = None;
@@ -319,7 +318,7 @@ impl CacheLifecycle {
             if state.ordinary_deadline.is_none() {
                 state.ordinary_deadline = Some(now + self.timeouts.ordinary);
             }
-            if self.build_cache.initial_store_pending() && state.initial_store_deadline.is_none() {
+            if self.cache.initial_store_pending() && state.initial_store_deadline.is_none() {
                 state.initial_store_deadline = Some(now + self.timeouts.initial_store);
             }
             let (reason, deadline) = [
@@ -376,7 +375,7 @@ impl CacheLifecycle {
             }
         }
         self.notify_changed();
-        let Some(current_target) = self.build_cache.pending_generation() else {
+        let Some(current_target) = self.cache.pending_generation() else {
             return;
         };
         {
@@ -397,15 +396,14 @@ impl CacheLifecycle {
     }
 
     async fn perform_flush(self: &Arc<Self>, target_generation: u64) {
-        let build_cache = self.build_cache.clone();
-        let result = match tokio::task::spawn_blocking(move || {
-            build_cache.publish_generation(target_generation)
-        })
-        .await
-        {
-            Ok(result) => result.map_err(|error| error.to_string()),
-            Err(error) => Err(format!("cache publication task failed: {error}")),
-        };
+        let cache = self.cache.clone();
+        let result =
+            match tokio::task::spawn_blocking(move || cache.publish_generation(target_generation))
+                .await
+            {
+                Ok(result) => result.map_err(|error| error.to_string()),
+                Err(error) => Err(format!("cache publication task failed: {error}")),
+            };
         let mut state = self
             .state
             .lock()
@@ -418,7 +416,7 @@ impl CacheLifecycle {
         }
         let should_schedule = match result {
             Ok(()) => {
-                if self.build_cache.pending_generation().is_none() {
+                if self.cache.pending_generation().is_none() {
                     state.initial_store_deadline = None;
                     state.ordinary_deadline = None;
                     state.large_change_deadline = None;
@@ -470,7 +468,7 @@ impl CacheLifecycle {
                     Action::Wait
                 } else {
                     drop(state);
-                    let target = self.build_cache.pending_generation();
+                    let target = self.cache.pending_generation();
                     let mut state = self
                         .state
                         .lock()
@@ -507,7 +505,7 @@ impl CacheLifecycle {
                 Action::Done(mut outcome) => {
                     outcome
                         .infrastructure_log_events
-                        .extend(self.build_cache.take_infrastructure_log_events());
+                        .extend(self.cache.take_infrastructure_log_events());
                     self.notify_changed();
                     return outcome;
                 }
@@ -557,7 +555,7 @@ impl CacheLifecycle {
                         } else {
                             let diagnostic = state.diagnostic.take();
                             drop(state);
-                            let target = self.build_cache.pending_generation();
+                            let target = self.cache.pending_generation();
                             let mut state = self
                                 .state
                                 .lock()
@@ -604,7 +602,7 @@ impl CacheLifecycle {
                 Action::Done(mut outcome) => {
                     outcome
                         .infrastructure_log_events
-                        .extend(self.build_cache.take_infrastructure_log_events());
+                        .extend(self.cache.take_infrastructure_log_events());
                     self.notify_changed();
                     return outcome;
                 }
@@ -627,7 +625,7 @@ impl CacheLifecycle {
                 diagnostic.is_none(),
                 "cache publication failed: {diagnostic:?}"
             );
-            if self.build_cache.pending_generation().is_none() && !flush_in_flight {
+            if self.cache.pending_generation().is_none() && !flush_in_flight {
                 return;
             }
             changed
@@ -661,8 +659,8 @@ impl Drop for CacheRunActivity {
 
 impl Compiler {
     pub fn new(options: CompilerOptions) -> Self {
-        let build_cache = BuildCache::new(options.cache.clone(), options.snapshot.clone());
-        let cache_lifecycle = CacheLifecycle::new(build_cache.clone(), &options.cache);
+        let cache = Cache::new(options.cache.clone(), options.snapshot.clone());
+        let cache_lifecycle = CacheLifecycle::new(cache.clone(), &options.cache);
         let mut hooks = CompilerHookSet::default();
         if options.provided_exports {
             FlagDependencyExportsPlugin.apply(&mut hooks);
@@ -677,7 +675,7 @@ impl Compiler {
         }
         Self {
             options,
-            build_cache,
+            cache,
             cache_lifecycle,
             hooks,
         }
@@ -693,7 +691,7 @@ impl Compiler {
         Compilation::new(
             self.options.clone(),
             UnpackResolver::new(self.options.resolve.clone()),
-            self.build_cache.clone(),
+            self.cache.clone(),
             compilation_hooks,
         )
     }
@@ -709,7 +707,7 @@ impl Compiler {
         &self,
         idle_reason: CacheIdleReason,
     ) -> Result<PendingCompilation> {
-        self.build_cache
+        self.cache
             .prepare_for_compilation(
                 &self.options.context,
                 &UnpackResolver::new(self.options.resolve.clone()),
@@ -731,14 +729,13 @@ impl Compiler {
         .instrument(tracing::trace_span!("Compiler::run"))
         .await;
         if result.is_ok() {
-            self.build_cache.store_build_dependencies();
-            self.build_cache.on_compilation_completed();
+            self.cache.store_build_dependencies();
+            self.cache.on_compilation_completed();
         }
-        self.build_cache.trace_work_counters();
+        self.cache.trace_work_counters();
         result.map(|mut compilation| {
-            compilation.extend_infrastructure_log_events(
-                self.build_cache.take_infrastructure_log_events(),
-            );
+            compilation
+                .extend_infrastructure_log_events(self.cache.take_infrastructure_log_events());
             PendingCompilation {
                 compilation: Some(compilation),
                 cache_activity: Some(cache_activity),
@@ -749,7 +746,7 @@ impl Compiler {
     pub fn flush_cache(&self) -> std::result::Result<(), String> {
         let span = tracing::trace_span!("Compiler::flush_cache");
         let _enter = span.enter();
-        self.build_cache
+        self.cache
             .flush_to_filesystem()
             .map_err(|error| error.to_string())
     }
@@ -826,14 +823,14 @@ mod tests {
         ));
 
         let first = compiler.run().await?;
-        let first_cache = compiler.build_cache.stats();
+        let first_cache = compiler.cache.stats();
         assert_eq!(first_cache.module_entries, 2);
         assert_eq!(first_cache.module_hits, 0);
         assert_eq!(first_cache.module_misses, 2);
         assert_eq!(first_cache.resolve_entries, 2);
         assert_eq!(first_cache.resolve_hits, 0);
         assert_eq!(first_cache.resolve_misses, 2);
-        let first_work = compiler.build_cache.work_counters();
+        let first_work = compiler.cache.work_counters();
         assert_eq!(
             first_work.for_family(CacheItemFamily::Resolve),
             CacheItemWork {
@@ -856,14 +853,14 @@ mod tests {
         );
 
         let second = compiler.run().await?;
-        let second_cache = compiler.build_cache.stats();
+        let second_cache = compiler.cache.stats();
         assert_eq!(second_cache.module_entries, 2);
         assert_eq!(second_cache.module_hits, 2);
         assert_eq!(second_cache.module_misses, 2);
         assert_eq!(second_cache.resolve_entries, 2);
         assert_eq!(second_cache.resolve_hits, 2);
         assert_eq!(second_cache.resolve_misses, 2);
-        let second_work = compiler.build_cache.work_counters();
+        let second_work = compiler.cache.work_counters();
         assert_eq!(
             second_work.for_family(CacheItemFamily::Resolve),
             CacheItemWork {
@@ -969,11 +966,11 @@ mod tests {
         assert!(!cache_location.join("packs/modules.cbor").exists());
 
         let second_compiler = Compiler::new(options);
-        assert_eq!(second_compiler.build_cache.stats().resolve_entries, 0);
-        assert_eq!(second_compiler.build_cache.stats().module_entries, 0);
+        assert_eq!(second_compiler.cache.stats().resolve_entries, 0);
+        assert_eq!(second_compiler.cache.stats().module_entries, 0);
 
         let second = second_compiler.run().await?;
-        let second_cache = second_compiler.build_cache.stats();
+        let second_cache = second_compiler.cache.stats();
         assert_eq!(second_cache.resolve_hits, 2);
         assert_eq!(second_cache.module_hits, 2);
         assert_eq!(asset_sources(&first), asset_sources(&second));
@@ -982,7 +979,7 @@ mod tests {
             second.module_graph().modules().as_ptr(),
             "a later Compiler must assemble a fresh ModuleGraph"
         );
-        let restored_work = second_compiler.build_cache.work_counters();
+        let restored_work = second_compiler.cache.work_counters();
         assert_eq!(
             restored_work.for_family(CacheItemFamily::Resolve),
             CacheItemWork {
@@ -1015,7 +1012,7 @@ mod tests {
         );
 
         let third = second_compiler.run().await?;
-        let repopulated_work = second_compiler.build_cache.work_counters();
+        let repopulated_work = second_compiler.cache.work_counters();
         assert_eq!(
             repopulated_work.for_family(CacheItemFamily::Resolve),
             CacheItemWork {
@@ -1057,7 +1054,7 @@ mod tests {
         );
         for restored_module in second.module_graph().modules() {
             let cached_record = second_compiler
-                .build_cache
+                .cache
                 .module_builds()
                 .get(restored_module.identity(), None)
                 .expect("restored Module Build Record should remain in Memory Cache");
@@ -1231,7 +1228,7 @@ mod tests {
         let entered = Arc::new(Barrier::new(2));
         let release = Arc::new(Barrier::new(2));
         compiler
-            .build_cache
+            .cache
             .install_publish_barrier(entered.clone(), release.clone());
 
         let settling_compiler = compiler.clone();
@@ -1285,7 +1282,7 @@ mod tests {
         second.run().await?;
         assert_eq!(
             second
-                .build_cache
+                .cache
                 .work_counters()
                 .for_family(CacheItemFamily::ModuleBuild),
             CacheItemWork {
@@ -1343,7 +1340,7 @@ mod tests {
         );
         assert_eq!(
             compiler
-                .build_cache
+                .cache
                 .work_counters()
                 .for_family(CacheItemFamily::ModuleBuild)
                 .evictions,
@@ -1358,7 +1355,7 @@ mod tests {
             "#,
         )?;
         let third = compiler.run().await?;
-        let work = compiler.build_cache.work_counters();
+        let work = compiler.cache.work_counters();
 
         assert!(
             asset_sources(&third)
@@ -1402,7 +1399,7 @@ mod tests {
         let first = compiler.run().await?;
         compiler.flush_cache()?;
         let second = compiler.run().await?;
-        let work = compiler.build_cache.work_counters();
+        let work = compiler.cache.work_counters();
 
         assert_eq!(asset_sources(&first), asset_sources(&second));
         assert_eq!(
@@ -1425,7 +1422,7 @@ mod tests {
                 evictions: 0,
             }
         );
-        assert_eq!(compiler.build_cache.stats().module_entries, 0);
+        assert_eq!(compiler.cache.stats().module_entries, 0);
         assert_ne!(
             first.module_graph().modules().as_ptr(),
             second.module_graph().modules().as_ptr()
@@ -1468,11 +1465,11 @@ mod tests {
         readonly_options.cache.readonly = true;
         readonly_options.cache.max_age = std::time::Duration::ZERO;
         let readonly_compiler = Compiler::new(readonly_options);
-        assert_eq!(readonly_compiler.build_cache.stats().resolve_entries, 0);
-        assert_eq!(readonly_compiler.build_cache.stats().module_entries, 0);
+        assert_eq!(readonly_compiler.cache.stats().resolve_entries, 0);
+        assert_eq!(readonly_compiler.cache.stats().module_entries, 0);
 
         let second = readonly_compiler.run().await?;
-        let readonly_cache = readonly_compiler.build_cache.stats();
+        let readonly_cache = readonly_compiler.cache.stats();
         assert_eq!(readonly_cache.resolve_entries, 2);
         assert_eq!(readonly_cache.module_entries, 2);
         readonly_compiler.flush_cache()?;
@@ -1532,7 +1529,7 @@ mod tests {
         warm_compiler.run().await?;
         assert_eq!(
             warm_compiler
-                .build_cache
+                .cache
                 .work_counters()
                 .for_family(CacheItemFamily::ModuleBuild),
             CacheItemWork {
@@ -1549,7 +1546,7 @@ mod tests {
         cold_compiler.run().await?;
         assert_eq!(
             cold_compiler
-                .build_cache
+                .cache
                 .work_counters()
                 .for_family(CacheItemFamily::ModuleBuild),
             CacheItemWork {
