+++ crates/unpack_core/src/make.rs
@@ -20,7 +20,7 @@ use crate::{
     AsyncDependenciesBlockIndex, CompilerOptions, Dependency, DependencyIndex, DependencyKind,
     Error, FactorizedModule, LoaderRequest, LoaderRunner, MatchedLoader, ModuleGraph, ModuleHandle,
     ModuleIdentity, NormalModuleFactory, Result, SnapshotStrategy, UnpackResolver,
-    cache::{BuildCache, ModuleBuildRecord},
+    cache::{Cache, ModuleBuildRecord},
     cache_facade::ModuleBuildCache,
     module::BuiltModuleContent,
     parser::{ParsedModule, parse_module_dependencies},
@@ -187,7 +187,7 @@ type BackgroundMakeTask = JoinHandle<Result<Vec<MakeTask>>>;
 pub(crate) async fn run(
     options: &CompilerOptions,
     resolver: UnpackResolver,
-    build_cache: BuildCache,
+    cache: Cache,
     file_system_info: FileSystemInfo,
     state: Arc<Mutex<MakeState>>,
 ) -> Result<()> {
@@ -195,14 +195,14 @@ pub(crate) async fn run(
     let services = MakeServices {
         normal_module_factory: NormalModuleFactory::new(
             resolver,
-            build_cache.normal_module_factory(),
+            cache.normal_module_factory(),
             file_system_info.clone(),
             options.snapshot.resolve,
             snapshot_cache.clone(),
         )
         .with_module_rules(options.module_rules.clone())
         .with_side_effects(options.side_effects != crate::SideEffectsOption::Disabled),
-        module_build_cache: build_cache.module_builds(),
+        module_build_cache: cache.module_builds(),
         file_system_info,
         module_snapshot_strategy: options.snapshot.module,
         snapshot_cache,
@@ -823,19 +823,19 @@ mod tests {
 
         let options = CompilerOptions::new(temp.path(), vec![Entry::new("main", "./index")]);
         let resolver = UnpackResolver::new(options.resolve.clone());
-        let build_cache = BuildCache::new(options.cache.clone(), options.snapshot.clone());
+        let cache = Cache::new(options.cache.clone(), options.snapshot.clone());
         let state = Arc::new(Mutex::new(MakeState::default()));
 
         run(
             &options,
             resolver,
-            build_cache.clone(),
+            cache.clone(),
             FileSystemInfo::new(),
             Arc::clone(&state),
         )
         .await?;
 
-        let cache = build_cache.stats();
+        let cache = cache.stats();
         let state = state.lock().await;
         let graph = &state.module_graph;
         let dep = graph
