+++ docs/adr/0131-use-rust-native-webpack-cache-layers-and-item-identities.md
@@ -1,3 +1,3 @@
 # Use Rust-native webpack cache layers and item identities
 
-Unpack will model webpack's Cache, Cache Facade, Memory Cache, idle and shutdown lifecycle, and Persistent Cache responsibilities with Rust-native interfaces owned by the Compiler, using Cache Namespace, Cache Identifier, and optional Cache ETag identity while keeping each Compilation and its graphs fresh. The shared Cache Layer seam is limited to lookup, store, and clear data-plane operations: Memory Cache owns generation retention, Persistent Cache owns preparation, access touches, restoration, and publication, and PackFile remains the lower-level Persistent Cache storage implementation. The top-level `cache` and `cache_facade` modules correspond to webpack's root `Cache.js` and `CacheFacade.js`; ordinary and generation-collected memory layers map separately to webpack's Memory Cache Plugin and Memory With GC Cache Plugin, while the Pack File Cache Strategy lives in the `cache` category. Compiler-owned `BuildCache`, Cache Items, options, and Pack File storage remain private Rust composition or storage helpers colocated with those webpack responsibilities rather than defining a separate `build_cache` architecture. Resolve Records, Module Build Records, Code Generation Records, and Asset Render Records remain independent Cache Items that validate their relevant inputs; resolved Build Dependency tracking includes the Unpack toolchain and supersedes ADR 0092's explicit-file approximation, extending ADR 0064, ADR 0085, and ADR 0095 without introducing webpack's Tapable mechanism or a whole-project Snapshot or Compiler-options fingerprint.
+Unpack will model webpack's Cache, Cache Facade, Memory Cache, idle and shutdown lifecycle, and Persistent Cache responsibilities with Rust-native interfaces owned by the Compiler, using Cache Namespace, Cache Identifier, and optional Cache ETag identity while keeping each Compilation and its graphs fresh. The shared Cache Layer seam is limited to lookup, store, and clear data-plane operations: Memory Cache owns generation retention, Persistent Cache owns preparation, access touches, restoration, and publication, and PackFile remains the lower-level Persistent Cache storage implementation. The top-level `cache` and `cache_facade` modules correspond to webpack's root `Cache.js` and `CacheFacade.js`; ordinary and generation-collected memory layers map separately to webpack's Memory Cache Plugin and Memory With GC Cache Plugin, while the Pack File Cache Strategy lives in the `cache` category. The Compiler owns the single deep `Cache` module, which hides lifecycle, facade composition, snapshots, diagnostics, and a private `CacheLayers` coordinator; there is no separate `BuildCache` type or architecture. Cache Items, options, and Pack File storage remain private Rust helpers colocated with those webpack responsibilities. Resolve Records, Module Build Records, Code Generation Records, and Asset Render Records remain independent Cache Items that validate their relevant inputs; resolved Build Dependency tracking includes the Unpack toolchain and supersedes ADR 0092's explicit-file approximation, extending ADR 0064, ADR 0085, and ADR 0095 without introducing webpack's Tapable mechanism or a whole-project Snapshot or Compiler-options fingerprint.
